# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/devices/{device_id}/services:
    get:
      operationId: listDeviceServices
      tags: [runtime]
      security: [{ adminBearer: [] }]
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: startup_type, in: query, schema: { type: string } }
        - { name: present, in: query, schema: { type: string, enum: ["true", "false", all] } }
      responses: { "200": { description: Paginated services, content: { application/json: { schema: { $ref: "#/components/schemas/PaginatedInventory" } } } }, "401": { description: Unauthorized }, "404": { description: Device not found } }
  /v1/devices/{device_id}/service-events:
    get:
      operationId: listDeviceServiceEvents
      tags: [runtime]
      security: [{ adminBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/DeviceId" }]
      responses: { "200": { description: Paginated service events, content: { application/json: { schema: { $ref: "#/components/schemas/PaginatedInventory" } } } } }
  /v1/devices/{device_id}/processes:
    get:
      operationId: listDeviceProcesses
      tags: [runtime]
      security: [{ adminBearer: [] }]
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: username, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string, enum: [process_name, cpu_percent, working_set_bytes, started_at] } }
      responses: { "200": { description: Current processes, content: { application/json: { schema: { $ref: "#/components/schemas/PaginatedInventory" } } } }, "401": { description: Unauthorized }, "404": { description: Device not found } }
  /v1/devices/{device_id}/processes/summary:
    get:
      operationId: getDeviceProcessesSummary
      tags: [runtime]
      security: [{ adminBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/DeviceId" }]
      responses: { "200": { description: Process summary, content: { application/json: { schema: { type: object, additionalProperties: true } } } } }
components:
  parameters:
    DeviceId:
      name: device_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    agentBearer:
      type: http
      scheme: bearer
    adminBearer:
      type: http
      scheme: bearer
    # Interactive console authentication. Deliberately separate from adminBearer,
    # which is a platform-level machine credential and is never issued to a tenant.
    sessionCookie:
      type: apiKey
      in: cookie
      name: nexora_session
  schemas:
...
    PaginatedInventory:
      type: object
      required: [items, page, page_size, total]
      properties:
        items: { type: array, items: { type: object, additionalProperties: true } }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
...